Extract wave service errors
